% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{color2hcl}
\alias{color2hcl}
\title{Convert named colors to HCL (polarLUV)}
\usage{
color2hcl(color)
}
\arguments{
\item{palette}{Character vector. A palette (vector of hex RGB or RGBA colors).}
}
\value{
Data frame with three columns: \code{$H}, \code{$C}, \code{$L}
}
\description{
Converts a hex color palette to HCL. Returns a dataframe. NA's in palette are preserved.
}
