% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{color2hex}
\alias{color2hex}
\title{Convert named colors to hex RGB}
\usage{
color2hex(color)
}
\arguments{
\item{color}{Character vector. A palette (vector of named colors).}
}
\value{
Character vector. A palette (vector of hex RGB colors).
}
\description{
Converts named colors to hex RGB colors. Returns a character vector of hex RGB colors.
}
