% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{colorblind}
\alias{colorblind}
\title{Simulate colorblindness}
\usage{
colorblind(palette)
}
\arguments{
\item{palette}{Character vector. A palette (vector of hex RGB or RGBA colors).}
}
\value{
List of four hex color palettes: \code{$normal}, \code{$deuteranopia}, \code{$protanopia}, \code{$tritanopia}
}
\description{
From color palette (character vector of hex RGB or RGBA color codes), simulates the palette as it
would appear under three common forms of colorblindness: deuteranopia, protanopia and tritanopia.
Returns a list of palettes. NA's in palette are preserved.
}
