% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{colorchart}
\alias{colorchart}
\title{Get built-in color names with Hex, RGB and HCL values}
\usage{
colorchart(distinct = FALSE)
}
\arguments{
\item{distinct}{logical indicating if the colors returned should all be distinct;
e.g., "snow" and "snow1" are synonyms with the same RGB values.}
}
\value{
Data frame with 1 row per color and 7 variables:
\item{hex}{Hexidecimal RGB color string}
\item{red}{Red value (0 to 255)}
\item{green}{Green value (0 to 255)}
\item{blue}{Blue value (0 to 255)}
\item{L}{Lightness value (0.0 to 100.0)}
\item{C}{Chroma value (0.0 to ~150.0). Note: Maximum C depends on L and H.}
\item{H}{Hue angle (0.0 to 360.0)}
}
\description{
Works like \code{grDevices::colors()} but returns a dataframe of Hex RGB codes, integer red, green, and blue
values, and the corresponding H, C, and L values (in HCL color space) for all color names.
}
