% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclx.R
\name{cyclx}
\alias{cyclx}
\alias{rampx}
\alias{stepx}
\title{Generate Power Function, Stepped, and Cyclic Sequences}
\usage{
cyclx(from = 0, to = 1, n = 2, step.n = 1, exponent = 1,
  cyc.from = NULL, cyc.to = NULL, cyc.n = n, cyc.exponent = 1,
  operator = "*")

rampx(from = 0, to = 1, n = 2, exponent = 1)

stepx(from = 0, to = 1, n = 2, step.n = 1, exponent = 1)
}
\arguments{
\item{from}{Numeric. Number to interpolate from. For \code{cyclx()}, first number in the trend.}

\item{to}{Numeric. Number to interpolate to. For \code{cyclx()}, last number in the trend.}

\item{n}{Numeric. Length of sequence. Equvalent of \code{length.out} in \code{\link[base:seq]{base::seq()}}.}

\item{step.n}{Numeric. Step length, in same units as \code{n}.}

\item{exponent}{Numeric. Exponent of power function used to interpolate betwwen \code{from} and \code{to}.}

\item{cyc.from}{Numeric. Number to interpolate from, for each cycle to be imposed on trend.}

\item{cyc.to}{Numeric. Number to interpolate to, for each cycle to be imposed on trend.}

\item{cyc.n}{Numeric. Length of cycle, in same units as \code{n}.}

\item{cyc.exponent}{Numeric. Exponent of power function used to interpolate betwwen \code{cyc.from}
and \code{cyc.to}.}

\item{operator}{Character. \code{"*"} or \code{"+"}. Operator used to impose cycles on trend. By default,
cycles and trend are multiplied ("*").}
}
\value{
Numeric vector
}
\description{
Generate Power Function, Stepped, and Cyclic Sequences
}
\examples{
s <- rampx(0,1,101)
stepped_s <- stepx(0,1,101,step.n = 10)
plot(s,type="l")
points(stepped_s,type="p",pch=20)

s <- rampx(0,1,101,exp=2)
stepped_s <- stepx(0,1,101,step.n = 20,exp=2)
plot(s,type="l")
points(stepped_s,type="p",pch=20)

s <- stepx(0,1,101,step.n=20)
cycled_s <- cyclx(0,1,101,step.n = 20,cyc.from=0.5,cyc.to=1,cyc.n=20,cyc.exponent = 1)
plot(cycled_s,type="p",pch=19,col="red",ylim=c(0,1))
points(s,type="p",pch=20)

stepped_s <- stepx(0,1,101,step.n = 20,exp=.5)
cycled_s <- cyclx(0,1,101,step.n = 20,exp=.5,cyc.from=0,cyc.to=1,cyc.n=20,cyc.exponent=1)
plot(cycled_s,type="p",pch=19,col="red")
points(stepped_s,type="p",pch=20)

stepped_s <- stepx(0,1,101,step.n=10,exp=2)
cycled_s <- cyclx(0,1,101,step.n=10,exp=2,cyc.from=0,cyc.to=1,cyc.n=10,cyc.exponent=2,operator="+")
plot(cycled_s,type="p",pch=19,col="red")
points(stepped_s,type="p",pch=20)

}
