% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{hcl2hex}
\alias{hcl2hex}
\title{Convert an HCL palette to hex RGB}
\usage{
hcl2hex(hcl = NULL, H = NULL, C = NULL, L = NULL)
}
\arguments{
\item{hcl}{Optional data frame with three columns: \code{$H}, \code{$C}, \code{$L}. (default = \code{NULL})}

\item{H}{Numeric. Hue angle (mod 360).}

\item{C}{Numeric. Chroma.  Minimum value = 0.  Maximum possible depends on Hue and Lightness (about 100 to 150)}

\item{H}{Numeric. Lightness (0 to 100)}
}
\value{
Character vector. A palette (vector of hex RGB colors).
}
\description{
Converts an HCL color palette supplied as a dataframe or individual arguments to an RGB palette.
Returns a character vector of hex RGB colors.
}
