% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{heat}
\alias{heat}
\title{A sequential color palette for heatmaps.}
\usage{
heat(n = 10)
}
\arguments{
\item{n}{Number of levels in palette (number of bins).  Default: \code{n = 10}.}
}
\value{
character vector. A vector of RGB colors specified as hex color codes (#RRGGBB).
}
\description{
This function returns a sequential red-to-yellow palette of length \code{n}, suitable as a replacement
for \link[grDevices:heat.colors]{grDevices::heat.colors} or \link[colorspace:heat_hcl]{colorspace::heat_hcl}.
}
\details{
Palette characteristics:
\itemize{
\item perceptually linear
\item color-blind-safe (under three most common forms of colorblindness) and print friendly
\item under normal vision: linear lightness, chroma, and hue gradients
\item under colorblind viewing, linear lightness, smooth chroma, flat hue gradients
}
}
\examples{
colorplot(heat(128))
}
