% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unipalette.R
\name{unipalette}
\alias{unipalette}
\title{A 7 category qualitative palette}
\usage{
unipalette(colors = NULL, order = "difference")
}
\arguments{
\item{colors}{Integer vector or character vector. \code{colors=4} returns the first 4 colors of the
default ordering. \code{colors=2:5} returns the colors 2 through 5 of the default ordering.
\code{colors=c("red","lightblue","purple")} returns the named colors in the order given.}

\item{order}{Character. \code{"difference"}, \code{"lightness"}, or \code{"hybrid"}. The ordering for the
palette (see details).}
}
\value{
character vector. A vector of named RGB colors specified as hex color codes (#RRGGBB).
}
\description{
\code{unipalette()} produces qualitative palettes based on a preset 7-color palette optimized for
lines, points, bars, etc, on a white background.
}
\details{
Base palette, in default ordering:

#' \tabular{lllllll}{
red \tab lightblue \tab black \tab green \tab purple \tab pink \tab blue \cr
"#a11c3e"\tab"#5798d1"\tab"#252525"\tab"#319045"\tab"#5e2b7b"\tab"#e2908c"\tab"#226e83"
}

Palette characteristics:
\itemize{
\item designed for high contrast against a white background, with good distinction among colors
\item colors distinguishable under two most common forms of colorblindness
\item colors distinguishable in greyscale for b/w printing
\item Good for bar charts, box plots, pies, etc.
\item Line should be relatively thick
\item Points should be relatively large
}

The default ordering (\code{order="difference"}) maximizes differences in hue and lightness between adjacent
colors. This aids discrimination of data plotted side by side, and also helps avoid implications
that the data is ordinal.

Two alternate orderings are also provided.
\itemize{
\item \code{unipalette(order="lightness")} orders the color by increasing perceived lightness:
black    purple       red      blue     green lightblue      pink
\item \code{unipalette(order="hybrid")} increases the pairwise lightness difference in an otherwise
lightness-ordered palette:
black       red    purple     green      blue      pink lightblue
}

Palette subsets

\code{unipalette(4)} returns the first 4 colors of the default ordering.
\code{unipalette(2:5)} returns the colors 2 through 5 of the default ordering.
\code{unipalette(c("red","lightblue","purple"))} returns the named colors in the order given.
}
\examples{
## 3 orderings:
difference <- unipalette(); colorbar(diff,dots=TRUE)
lightness <- unipalette(order="lightness"); colorbar(lightness,dots=TRUE)
hybrid <- unipalette(order="hybrid"); colorbar(hybrid,dots=TRUE)

## subsets
p <- unipalette(4); colorbar(p,dots=TRUE)
}
