% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vpad}
\alias{vpad}
\title{pad a vector}
\usage{
vpad(x, n = length(x) + 1, pad = x[length(x)])
}
\arguments{
\item{x}{A vector}

\item{n}{Integer. New length for vector.}

\item{pad}{Vector. Element or vector with which to pad original vector. Must be a multiple of n.}
}
\value{
vector
}
\description{
pad a vector
}
