% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar.R
\name{colorbar}
\alias{colorbar}
\title{Display a palette as a color bar.}
\usage{
colorbar(p)
}
\arguments{
\item{p}{A palette (vector of colors).}
}
\description{
This function plots a palette as a horizontal colorbar.
}
